+++ lib/shared/common.ts
@@ -1233,6 +1233,7 @@ let handlePlugins = async (
           if (namespace != null) request.namespace = namespace
           if (resolveDir != null) request.resolveDir = resolveDir
           if (kind != null) request.kind = kind
+          else throw new Error(`Must specify "kind" when calling "resolve"`)
           if (pluginData != null) request.pluginData = details.store(pluginData)
 
           sendRequest<protocol.ResolveRequest, protocol.ResolveResponse>(refs, request, (error, response) => {