+++ pkg/api/api.go
@@ -577,7 +577,8 @@ type OnLoadResult struct {
 type ResolveKind uint8
 
 const (
-	ResolveEntryPoint ResolveKind = iota
+	ResolveNone ResolveKind = iota
+	ResolveEntryPoint
 	ResolveJSImportStatement
 	ResolveJSRequireCall
 	ResolveJSDynamicImport